--- src/main/java/redlightBack/Board/BoardRepository.java
@@ -1,6 +1,0 @@
-package redlightBack.Board;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BoardRepository extends JpaRepository<Board, Long> {
-}
