--- src/main/java/redlightBack/Board/Dto/BoardResponse.java
@@ -1,9 +1,0 @@
-package redlightBack.Board.Dto;
-
-import java.time.LocalDateTime;
-
-public record BoardResponse(Long id,
-                            String boardName,
-                            int postCount
-                            ) {
-}
