--- src/main/java/redlightBack/Board/Dto/CreateBoardRequest.java
@@ -1,5 +1,0 @@
-package redlightBack.Board.Dto;
-
-public record CreateBoardRequest(String boardName
-                           ) {
-}
