--- src/main/java/redlightBack/Post/Dto/DeletePostResponse.java
@@ -1,6 +1,0 @@
-package redlightBack.Post.Dto;
-
-import java.time.LocalDateTime;
-
-public record DeletePostResponse(LocalDateTime deletedAt) {
-}
