--- src/main/java/redlightBack/Post/Dto/PageInfo.java
@@ -1,7 +1,0 @@
-package redlightBack.Post.Dto;
-
-public record PageInfo(int pageNumber,
-                       int size,
-                       long totalElements,
-                       int totalPages) {
-}
