--- src/main/java/redlightBack/Post/Dto/PostLikeResponse.java
@@ -1,6 +1,0 @@
-package redlightBack.Post.Dto;
-
-public record PostLikeResponse(Long postId,
-                               boolean liked,
-                               long likeCount) {
-}
