--- src/main/java/redlightBack/Post/Dto/PostListAndPagingResponse.java
@@ -1,8 +1,0 @@
-package redlightBack.Post.Dto;
-
-import java.util.List;
-
-public record PostListAndPagingResponse(String boardName,
-                                        List<PostListResponse> postListResponse,
-                                        PageInfo pageInfo) {
-}
