--- src/main/java/redlightBack/Post/Dto/PostListResponse.java
@@ -1,12 +1,0 @@
-package redlightBack.Post.Dto;
-
-import java.time.LocalDateTime;
-
-public record PostListResponse (Long id,
-                                String postTitle,
-                                String userId,
-                                int viewCount,
-                                int commentCount,
-                                long likeCount,
-                                LocalDateTime createdAt){
-}
