--- src/main/java/redlightBack/Post/Dto/PostResponse.java
@@ -1,17 +1,0 @@
-package redlightBack.Post.Dto;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public record PostResponse(Long boardId,
-                           Long id,
-                           String postTitle,
-                           String userId,
-                           int viewCount,
-                           String content,
-                           List<String> imageUrls,
-                           LocalDateTime createdAt,
-                           LocalDateTime updatedAt,
-                           long likeCount,
-                           int commentCount) {
-}
