--- src/main/java/redlightBack/Post/Dto/PostsLikedAndPagingResponse.java
@@ -1,7 +1,0 @@
-package redlightBack.Post.Dto;
-
-import java.util.List;
-
-public record PostsLikedAndPagingResponse(List<PostsLikedResponse> likedPostsResponse,
-                                          PageInfo pageInfo) {
-}
