--- src/main/java/redlightBack/Post/Dto/PostsLikedResponse.java
@@ -1,13 +1,0 @@
-package redlightBack.Post.Dto;
-
-import java.time.LocalDateTime;
-
-public record PostsLikedResponse(Long boardId,
-                                 String boardName,
-                                 String postTitle,
-                                 String userId,
-                                 int viewCount,
-                                 int commentCount,
-                                 long likeCount,
-                                 LocalDateTime createdAt) {
-}
