--- src/main/java/redlightBack/Post/Enum/Direction.java
@@ -1,6 +1,0 @@
-package redlightBack.Post.Enum;
-
-public enum Direction {
-    ASC, //오름차순
-    DESC //내림차순
-}
