--- src/main/java/redlightBack/Post/Enum/SortBy.java
@@ -1,6 +1,0 @@
-package redlightBack.Post.Enum;
-
-public enum SortBy {
-    NEWEST, //최신순 정렬
-    MOST_LIKED //좋아요순 정렬
-}
