--- src/main/java/redlightBack/Post/PostLikeRepository.java
@@ -1,8 +1,0 @@
-package redlightBack.Post;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
-
-    PostLike findByPostIdAndUserId(Long postId, String userId);
-}
