--- src/main/java/redlightBack/Post/PostRepository.java
@@ -1,9 +1,0 @@
-package redlightBack.Post;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-    Optional<Post> findByIdAndDeletedAtIsNull(Long postId);
-}
