+++ src/main/java/redlightBack/board/Board.java
@@ -1,0 +1,27 @@
+package redlightBack.board;
+
+
+import jakarta.persistence.*;
+import lombok.*;
+import redlightBack.common.BaseEntity;
+
+@Entity
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+public class Board extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String boardName;
+    int postCount;
+
+    public Board(String boardName) {
+        this.boardName = boardName;
+    }
+
+}
