+++ src/main/java/redlightBack/board/BoardRepository.java
@@ -1,0 +1,6 @@
+package redlightBack.board;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BoardRepository extends JpaRepository<Board, Long> {
+}
