+++ src/main/java/redlightBack/board/dto/BoardResponse.java
@@ -1,0 +1,7 @@
+package redlightBack.board.dto;
+
+public record BoardResponse(Long id,
+                            String boardName,
+                            int postCount
+                            ) {
+}
