+++ src/main/java/redlightBack/board/dto/CreateBoardRequest.java
@@ -1,0 +1,5 @@
+package redlightBack.board.dto;
+
+public record CreateBoardRequest(String boardName
+                           ) {
+}
