+++ src/main/java/redlightBack/post/PostLikeRepository.java
@@ -1,0 +1,8 @@
+package redlightBack.post;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
+
+    PostLike findByPostIdAndUserId(Long postId, String userId);
+}
