+++ src/main/java/redlightBack/post/PostRepository.java
@@ -1,0 +1,9 @@
+package redlightBack.post;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+    Optional<Post> findByIdAndDeletedAtIsNull(Long postId);
+}
