+++ src/main/java/redlightBack/post/dto/DeletePostResponse.java
@@ -1,0 +1,6 @@
+package redlightBack.post.dto;
+
+import java.time.LocalDateTime;
+
+public record DeletePostResponse(LocalDateTime deletedAt) {
+}
