+++ src/main/java/redlightBack/post/dto/PageInfo.java
@@ -1,0 +1,7 @@
+package redlightBack.post.dto;
+
+public record PageInfo(int pageNumber,
+                       int size,
+                       long totalElements,
+                       int totalPages) {
+}
