+++ src/main/java/redlightBack/post/dto/PostLikeResponse.java
@@ -1,0 +1,6 @@
+package redlightBack.post.dto;
+
+public record PostLikeResponse(Long postId,
+                               boolean liked,
+                               long likeCount) {
+}
