+++ src/main/java/redlightBack/post/dto/PostListAndPagingResponse.java
@@ -1,0 +1,8 @@
+package redlightBack.post.dto;
+
+import java.util.List;
+
+public record PostListAndPagingResponse(String boardName,
+                                        List<PostListResponse> postListResponse,
+                                        PageInfo pageInfo) {
+}
