+++ src/main/java/redlightBack/post/dto/PostsLikedAndPagingResponse.java
@@ -1,0 +1,7 @@
+package redlightBack.post.dto;
+
+import java.util.List;
+
+public record PostsLikedAndPagingResponse(List<PostsLikedResponse> likedPostsResponse,
+                                          PageInfo pageInfo) {
+}
