+++ src/main/java/redlightBack/post/dto/PostsLikedResponse.java
@@ -1,0 +1,13 @@
+package redlightBack.post.dto;
+
+import java.time.LocalDateTime;
+
+public record PostsLikedResponse(Long boardId,
+                                 String boardName,
+                                 String postTitle,
+                                 String userId,
+                                 int viewCount,
+                                 int commentCount,
+                                 long likeCount,
+                                 LocalDateTime createdAt) {
+}
