+++ src/main/java/redlightBack/post/enums/Direction.java
@@ -1,0 +1,6 @@
+package redlightBack.post.enums;
+
+public enum Direction {
+    ASC, //오름차순
+    DESC //내림차순
+}
