+++ src/main/java/redlightBack/post/enums/SortBy.java
@@ -1,0 +1,6 @@
+package redlightBack.post.enums;
+
+public enum SortBy {
+    NEWEST, //최신순 정렬
+    MOST_LIKED //좋아요순 정렬
+}
